sql: Move Psycopg2Backend to package backend
