--- Orange/data/sql/__init__.py
@@ -1,2 +1,0 @@
-import Orange.misc
-psycopg2 = Orange.misc.import_late_warning("psycopg2")
